requirements.txt
numpy
opencv-python-headless
pandas
Pillow==9.5.0
Pillow==9.5.0
Requests
scikit_learn
streamlit
tensorflow

